Move the code to libbeat
