+++ metricbeat/module/redis/test_redis.py
@@ -29,6 +29,7 @@ class Test(metricbeat.BaseTest):
     COMPOSE_SERVICES = ['redis']
 
     @unittest.skipUnless(metricbeat.INTEGRATION_TESTS, "integration test")
+    @metricbeat.tag('redis')
     @attr('integration')
     def test_info(self):
         """
